% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energypal.R
\name{hierarchy_carrier_palette}
\alias{hierarchy_carrier_palette}
\title{Get carrier color vector from hierarchy}
\usage{
hierarchy_carrier_palette(include_subtypes = FALSE)
}
\arguments{
\item{include_subtypes}{logical include subtype colors too}
}
\value{
named character vector
}
\description{
Get carrier color vector from hierarchy
}
