% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energypal.R
\name{load_energy_palettes}
\alias{load_energy_palettes}
\title{Load hierarchical energy palette specification}
\usage{
load_energy_palettes(force = FALSE)
}
\arguments{
\item{force}{Logical; if TRUE re-read file even if cached}
}
\value{
list representing YAML structure
}
\description{
Reads and caches the hierarchical palette YAML (carriers + technologies).
}
